refactor: merge weekly birthday update into send_monthly_birthday_summary
Apply PR review feedback:
- Reuse the existing birthday_summary_email_recipients SiteSettings field
  instead of adding a new one; remove the new field and its migration
- Default rolling-window lookahead is now 7 days
- Generalize into a single management command: --days [N] on
  send_monthly_birthday_summary sends the rolling-window update (bare
  --days defaults to 7 days), while the default invocation keeps the
  calendar-month summary behavior; delete send_weekly_birthday_summary
- Merge the weekly command tests into test_birthday_command
